+++ src/components/Forms/SignInForm/SignInForm.js
@@ -3,15 +3,11 @@ import Avatar from '@material-ui/core/Avatar';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
-import Link from '@material-ui/core/Link';
-import Box from '@material-ui/core/Box';
 import { Redirect } from 'react-router-dom';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
+import useStyles from './style';
 import { serviceRequest } from '../../../services/serviceRequest';
 import config from '../../../config';
 import { ErrorMessage } from '../../ErrorMessage';
@@ -19,44 +15,15 @@ import { ErrorMessage } from '../../ErrorMessage';
 const errorMsg = 'Invalid username or password';
 const path = '/api/v1/signin';
 const domain = config.apiDomain;
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {'Copyright © '}
-      <Link color="inherit" href="https://material-ui.com/">
-        Your Website
-      </Link>
-      {' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-}
 
-const useStyles = makeStyles((theme) => ({
-  '@global': {
-    body: {
-      backgroundColor: theme.palette.common.white,
-    },
-  },
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main,
-  },
-  form: {
-    width: '100%',
-    marginTop: theme.spacing(1),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-}));
+function preparePayload(method, data) {
+  const url = `https://${domain}${path}`;
+  return {
+    method,
+    url,
+    data,
+  };
+}
 
 export const SignInForm = () => {
   const classes = useStyles();
@@ -68,15 +35,6 @@ export const SignInForm = () => {
     const updateForm = { ...formData };
     updateForm[event.target.name] = event.target.value;
     setForm(updateForm);
-  }
-
-  function preparePayload(method, data) {
-    const url = `https://${domain}${path}`;
-    return {
-      method,
-      url,
-      data,
-    };
   }
 
   async function handleSubmit(e) {
@@ -108,7 +66,7 @@ export const SignInForm = () => {
           Sign in
         </Typography>
         {isError && <ErrorMessage message={errorMsg} styles={{ color: 'red' }} />}
-        <form className={classes.form} noValidate onSubmit={handleSubmit}>
+        <form className={classes.form} onSubmit={handleSubmit}>
           <TextField
             variant="outlined"
             margin="normal"
@@ -130,12 +88,8 @@ export const SignInForm = () => {
             label="Password"
             type="password"
             id="password"
-            autoComplete="current-password"
+            autoComplete="password"
             onChange={handleChange}
-          />
-          <FormControlLabel
-            control={<Checkbox value="remember" color="primary" />}
-            label="Remember me"
           />
           <Button
             type="submit"
@@ -148,9 +102,6 @@ export const SignInForm = () => {
           </Button>
         </form>
       </div>
-      <Box mt={8}>
-        <Copyright />
-      </Box>
     </Container>
   );
 };
